refactor: extract isDuplicateMessage from SqsResponseProcessor to Bucket
